Make pre-commit checks pass and automatically trigger on CI

+++ emg2qwerty/metrics.py
@@ -9,9 +9,9 @@ from typing import Any
 
 import Levenshtein
 import torch
+from torchmetrics import Metric
 
 from emg2qwerty.data import LabelData
-from torchmetrics import Metric
 
 
 class CharacterErrorRates(Metric):
@@ -50,7 +50,7 @@ class CharacterErrorRates(Metric):
 
     def compute(self) -> dict[str, float]:
         def _error_rate(errors: torch.Tensor) -> float:
-            return errors.item() / self.target_len.item() * 100.0
+            return float(errors.item() / self.target_len.item() * 100.0)
 
         return {
             "CER": _error_rate(self.insertions + self.deletions + self.substitutions),
